+++ .eslintrc.js
@@ -30,5 +30,6 @@ module.exports = {
         "padded-blocks": 0,
         "func-names": 0,
         "max-len": 0,
+        "no-underscore-dangle": 0,
     }
 };