+++ src/controllers/image.js
@@ -4,11 +4,11 @@
  * En este archivo se recopila todo lo que se puede 
  * hacer con las imágenes
  */
-const path = require('path');
-const { randomNumber } = require('../helpers/libs');
-const fs = require('fs-extra');
-
-const { Image } = require('../models/index');
+const path = require('path');                             // Funciones para nombres de directorios
+const fs = require('fs-extra');                           // Funciones extra file system
+const md5 = require('md5');                               // Manejo de hash
+const { randomNumber } = require('../helpers/libs');      // Helper para nombres aleatorios
+const { Image, Comment } = require('../models/index');    // Modelos de la base de datos
 
 const ctrl = {};
 
@@ -18,9 +18,9 @@ const ctrl = {};
 ctrl.index = async (req, res) => {
   // Búsqueda. La expresión regular es para que incluya la extensión en la dirección 
   const image = await Image.findOne({ filename: { $regex: req.params.image_id } });  
-  res.render('image', { image });
+  const comments = await Comment.find({ image_id: image._id });
+  res.render('image', { image, comments });
 };
-
 
 /** Subiendo imagen al servidor 
  * POST /images
@@ -67,7 +67,22 @@ ctrl.like = (req, res) => {
 
 };
 
-ctrl.comment = (req, res) => {
+/** Añade comentario
+ * POST /images/:image_id/comment
+ * Usando los datos de la imagen y el formulario, crea el modelo del comentario
+ */
+ctrl.comment = async (req, res) => {
+  const image = await Image.findOne({ filename: { $regex: req.params.image_id } });
+  if (image) {
+    const newComment = new Comment(req.body);
+    newComment.gravatar = md5(newComment.email); // Gravatar requiere este hash para identificar al usuario
+    newComment.image_id = image._id;
+    console.log(newComment)
+    await newComment.save();
+    res.redirect(`/images/${image.uniqueId}#${newComment._id}`);
+  } else {
+    res.redirect('/');
+  }
 
 };