+++ src/models/index.js
@@ -1,8 +1,8 @@
-
+/** Este archivo redireciona los modelos indexados */
 
 
 module.exports = {
 
   Image: require('./Image'),
-
+  Comment: require('./Comment'),
 };